Skip package caching (saves a minute on Windows)
We'll try different combinations of setting the environment variable;
it's possible that we won't have to set it in every single step, but I
suspect we will have to set it in every step.

Note that once we go to .Net Core 3 exclusively, we'll be able to get
rid of this.